move queue-table-for-all-statuses into view/
