+++ packages/esm-service-queues-app/src/views/queue-tables-for-all-statuses.component.tsx
@@ -2,17 +2,17 @@ import { InlineNotification } from '@carbon/react';
 import { ExtensionSlot, isDesktop, useLayoutType } from '@openmrs/esm-framework';
 import React, { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import styles from './queue-table.scss';
+import styles from '../queue-table/queue-table.scss';
 import { useQueueEntries } from '../hooks/useQueueEntries';
 import type { QueueEntry, Queue, QueueTableTabConfig, QueueTableColumn } from '../types';
-import { queueTableComingFromColumn } from './cells/queue-table-coming-from-cell.component';
-import { queueTableNameColumn } from './cells/queue-table-name-cell.component';
-import { queueTablePriorityColumn } from './cells/queue-table-priority-cell.component';
-import { queueTableStatusColumn } from './cells/queue-table-status-cell.component';
-import { queueTableWaitTimeColumn } from './cells/queue-table-wait-time-cell.component';
-import { QueueTableByStatusSkeleton } from './queue-table-by-status-skeleton.component';
-import QueueTable from './queue-table.component';
-import { queueTableActionColumn } from './cells/queue-table-action-cell.component';
+import { queueTableComingFromColumn } from '../queue-table/cells/queue-table-coming-from-cell.component';
+import { queueTableNameColumn } from '../queue-table/cells/queue-table-name-cell.component';
+import { queueTablePriorityColumn } from '../queue-table/cells/queue-table-priority-cell.component';
+import { queueTableStatusColumn } from '../queue-table/cells/queue-table-status-cell.component';
+import { queueTableWaitTimeColumn } from '../queue-table/cells/queue-table-wait-time-cell.component';
+import { QueueTableByStatusSkeleton } from '../queue-table/queue-table-by-status-skeleton.component';
+import QueueTable from '../queue-table/queue-table.component';
+import { queueTableActionColumn } from '../queue-table/cells/queue-table-action-cell.component';
 import PatientSearch from '../patient-search/patient-search.component';
 import { Add } from '@carbon/react/icons';
 import { Search } from '@carbon/react';
